Restructure handlers. Add ui components for integrations.
